pyproject.toml
[project]
name = "rainbow-lunarlander"
version = "0.1.0"
description = "Modular RL experimentation platform with real-time visualization for Lunar Lander"
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
    "torch>=2.1.0",
    "gymnasium[box2d]>=0.29.0",
    "numpy>=1.24.0",
    "pyqt6>=6.6.0",
    "pyqtgraph>=0.13.3",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black",
    "ruff",
]

[project.scripts]
rl-gui = "main:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py310"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

# Explicitly configure PyTorch with CUDA 12.4
[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
